back/seo/requirements.txt
astroid==2.3.2
backcall==0.1.0
beautifulsoup4==4.8.1
bootstrap4==0.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-bootstrap==0.2.4
django-bootstrap4==1.0.1
django-cors-headers==3.2.1
django-extensions==2.2.5
django-heroku==0.3.1
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.10.3
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.0
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.8.4
pycparser==2.19
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.6
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.7
whitenoise==4.1.4
wrapt==1.11.2
